requirements.txt
# conda install ffmpeg -y && conda update ffmpeg -y

# For older GPU
# pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu124
# conda install -y -c rapidsai -c nvidia -c conda-forge cuml cuda-version=12.4
# conda install -y cuda-cudart cuda-version=12

# For newer GPU
# pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu130
# conda install -y -c rapidsai -c nvidia -c conda-forge cuml cuda-version=13.0
# conda install -y cuda-cudart cuda-version=13

pytorch-lightning
transformers
vector-quantize-pytorch
tensorboard
librosa
soundfile
ffmpeg
pyyaml
pyworld
praat-parselmouth
scikit-learn
tqdm
